refactor: 메인 페이지에서 핸들링하도록 변경
